chore: changed styles to make it responsive
